+++ app/src/main/java/com/strv/movies/model/Genre.kt
@@ -1,6 +1,8 @@
 package com.strv.movies.model
 
 import com.squareup.moshi.Json
+import com.strv.movies.data.entity.GenreEntity
+import com.strv.movies.data.entity.MovieGenreEntity
 
 data class Genre(
     val id: Int,
@@ -10,4 +12,14 @@ data class Genre(
 data class GenreDTO(
     @Json(name = "id") val id: Int,
     @Json(name = "name") val name: String
+)
+
+fun GenreDTO.toEntity() = GenreEntity(
+    genreId = id,
+    name = name
+)
+
+fun GenreDTO.toEntity(movieId: Int) = MovieGenreEntity(
+    genreId = id,
+    movieId = movieId
 )
