+++ app/src/main/java/com/strv/movies/model/MovieDetail.kt
@@ -1,6 +1,7 @@
 package com.strv.movies.model
 
 import com.squareup.moshi.Json
+import com.strv.movies.data.entity.MovieDetailEntity
 
 // Used for getting data from network
 data class MovieDetailDTO(
@@ -31,4 +32,14 @@ data class MovieDetail(
     val posterPath: String,
     val runtime: Int, // Not used for now - try to include it in UI if you want :)
     val genres: List<Genre>
+)
+
+fun MovieDetailDTO.toEntity() = MovieDetailEntity(
+    id = id,
+    title = title,
+    overview = overview,
+    releaseDate = releaseDate,
+    posterPath = posterPath,
+    runtime = runtime,
+    revenue = revenue
 )
