+++ app/src/main/java/com/strv/movies/network/MovieRepository.kt
@@ -2,15 +2,14 @@ package com.strv.movies.network
 
 import androidx.room.withTransaction
 import com.strv.movies.data.dao.MoviesDao
-import com.strv.movies.data.mapper.MovieDetailMapper
+import com.strv.movies.data.entity.toDomain
 import com.strv.movies.data.mapper.MovieMapper
-import com.strv.movies.data.mapper.toDomain
-import com.strv.movies.data.mapper.toEntity
 import com.strv.movies.database.MoviesDatabase
 import com.strv.movies.extension.Either
 import com.strv.movies.model.Movie
 import com.strv.movies.model.MovieDetail
 import com.strv.movies.model.MovieDetailDTO
+import com.strv.movies.model.toEntity
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
@@ -20,16 +19,15 @@ import javax.inject.Singleton
 class MovieRepository @Inject constructor(
     private val api: MovieApi,
     private val moviesDao: MoviesDao,
-    private val movieDetailMapper: MovieDetailMapper,
     private val movieMapper: MovieMapper,
     private val moviesDatabase: MoviesDatabase
 ) {
 
-    suspend fun getMovieDetail(movieId: Int): Either<String, MovieDetail> {
+    suspend fun fetchMovieDetail(movieId: Int): Either<String, String> {
         return try {
             val movie = api.getMovieDetail(movieId)
             storeMovieDetail(movie)
-            Either.Value(movieDetailMapper.map(movie))
+            Either.Value(movie.title)
         } catch (exception: Throwable) {
             Either.Error(exception.localizedMessage ?: "Network error")
         }
